requirements.txt
Django==3.2.*
django-cassandra-engine==1.6.3
djangorestframework==3.12.4
django-filter==21.1